IDEMPIERE-5391 Unit Test: Fix support for parallel execution of test
