+++ org.adempiere.base/src/org/compiere/model/MTable.java
@@ -24,7 +24,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -151,10 +150,9 @@ public class MTable extends X_AD_Table implements ImmutablePOSupport
 	{
 		if (tableName == null)
 			return null;
-		Iterator<MTable> it = s_cache.values().iterator();
-		while (it.hasNext())
+		MTable[] tables = s_cache.values().toArray(new MTable[0]);
+		for (MTable retValue : tables)
 		{
-			MTable retValue = it.next();
 			if (tableName.equalsIgnoreCase(retValue.getTableName()))
 			{
 				return s_cache.get (ctx, retValue.get_ID(), e -> new MTable(ctx, e));
