+++ org.idempiere.test/src/org/idempiere/test/model/InventoryTest.java
@@ -29,15 +29,12 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.math.BigDecimal;
 import java.sql.Timestamp;
-import java.util.Arrays;
 import java.util.Properties;
 
 import org.compiere.model.MAcctSchema;
-import org.compiere.model.MAttributeSet;
 import org.compiere.model.MAttributeSetInstance;
 import org.compiere.model.MBPartner;
 import org.compiere.model.MClient;
@@ -48,16 +45,10 @@ import org.compiere.model.MInventory;
 import org.compiere.model.MInventoryLine;
 import org.compiere.model.MOrder;
 import org.compiere.model.MOrderLine;
-import org.compiere.model.MPriceList;
-import org.compiere.model.MPriceListVersion;
 import org.compiere.model.MProduct;
-import org.compiere.model.MProductPrice;
-import org.compiere.model.MStorageOnHand;
-import org.compiere.model.MWarehouse;
 import org.compiere.process.DocAction;
 import org.compiere.process.DocumentEngine;
 import org.compiere.process.ProcessInfo;
-import org.compiere.util.CacheMgt;
 import org.compiere.util.Env;
 import org.compiere.util.TimeUtil;
 import org.compiere.wf.MWorkflow;
@@ -187,196 +178,6 @@ public class InventoryTest extends AbstractTestCase {
 		if (!receipt1.isPosted()) {
 			String error = DocumentEngine.postImmediate(Env.getCtx(), receipt1.getAD_Client_ID(), receipt1.get_Table_ID(), receipt1.get_ID(), false, getTrxName());
 			assertNull(error, error);
-		}
-	}
-	
-	@Test
-	public void testSkipProductWithSerial() {
-		Properties ctx = Env.getCtx();
-		String trxName = getTrxName();
-		
-		MAttributeSet set = new MAttributeSet(Env.getCtx(), DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id, null);
-		set.setIsSerNo(true);
-		set.saveEx();
-			
-		MWarehouse wh = new MWarehouse(Env.getCtx(), DictionaryIDs.M_Warehouse.HQ.id, null);
-		boolean disallow = wh.isDisallowNegativeInv(); 
-		MProduct product = null;
-		try {
-			
-			if (!disallow) {
-				wh.setIsDisallowNegativeInv(true);
-				wh.saveEx();
-				CacheMgt.get().reset(MWarehouse.Table_Name, wh.get_ID());
-			}
-			
-			product = new MProduct(ctx, 0, null);
-			product.setM_Product_Category_ID(DictionaryIDs.M_Product_Category.CHEMICALS.id);
-			product.setName("testSkipProductWithSerial");
-			product.setValue("testSkipProductWithSerial");
-			product.setProductType(MProduct.PRODUCTTYPE_Item);
-			product.setIsStocked(true);
-			product.setIsSold(true);
-			product.setIsPurchased(true);
-			product.setC_UOM_ID(DictionaryIDs.C_UOM.EACH.id);
-			product.setC_TaxCategory_ID(DictionaryIDs.C_TaxCategory.STANDARD.id);
-			product.setM_AttributeSet_ID(DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id);
-			product.saveEx();
-	
-			MPriceListVersion plv = MPriceList.get(DictionaryIDs.M_PriceList.PURCHASE.id).getPriceListVersion(null);
-			MProductPrice pp = new MProductPrice(Env.getCtx(), 0, getTrxName());
-			pp.setM_PriceList_Version_ID(plv.getM_PriceList_Version_ID());
-			pp.setM_Product_ID(product.get_ID());
-			pp.setPriceStd(new BigDecimal("2"));
-			pp.setPriceList(new BigDecimal("2"));
-			pp.saveEx();
-			
-			MAttributeSetInstance asi = new MAttributeSetInstance(Env.getCtx(), 0, getTrxName());
-			asi.setM_AttributeSet_ID(DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id);
-			asi.setSerNo("testSkipProductWithSerial #1");
-			asi.saveEx();
-			
-			createPOAndMRForProduct(product.get_ID(), asi);
-			
-			MStorageOnHand[] onhands = MStorageOnHand.getOfProduct(Env.getCtx(), product.get_ID(), getTrxName());
-			assertEquals(1, onhands.length, "Unexpected number of on hand records");
-			assertEquals(onhands[0].getM_AttributeSetInstance_ID(), asi.get_ID(), "Unexpected M_AttributeSetInstance_ID for on hand record");
-			
-			MInventory inventory = new MInventory(ctx, 0, trxName);
-			inventory.setM_Warehouse_ID(DictionaryIDs.M_Warehouse.HQ.id);
-			inventory.setC_DocType_ID(DictionaryIDs.C_DocType.MATERIAL_PHYSICAL_INVENTORY.id);
-			inventory.saveEx();
-	
-			MInventoryLine iline = new MInventoryLine(inventory,
-					DictionaryIDs.M_Locator.HQ.id, 
-					product.getM_Product_ID(),
-					0, // M_AttributeSetInstance_ID
-					Env.ONE, // QtyBook
-					Env.ZERO);
-			iline.saveEx();
-	
-			//should error out with negative on hand (skip the only asi record with serno)
-			ProcessInfo info = MWorkflow.runDocumentActionWorkflow(inventory, DocAction.ACTION_Complete);
-			assertTrue(info.isError(), info.getSummary());
-		} finally {
-			rollback();
-			set.setIsSerNo(false);
-			set.saveEx();
-			
-			if (product != null) 
-				product.deleteEx(true);
-			
-			if (!disallow) {
-				wh.setIsDisallowNegativeInv(false);
-				wh.saveEx();
-				CacheMgt.get().reset(MWarehouse.Table_Name, wh.get_ID());
-			}
-		}
-	}
-	
-	@Test
-	public void testSkipProductWithSerial2() {
-		Properties ctx = Env.getCtx();
-		String trxName = getTrxName();
-		
-		MAttributeSet set = new MAttributeSet(Env.getCtx(), DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id, null);
-		set.setIsSerNo(true);
-		set.saveEx();
-			
-		MWarehouse wh = new MWarehouse(Env.getCtx(), DictionaryIDs.M_Warehouse.HQ.id, null);
-		boolean disallow = wh.isDisallowNegativeInv(); 
-		MProduct product = null;
-		try {
-			
-			if (!disallow) {
-				wh.setIsDisallowNegativeInv(true);
-				wh.saveEx();
-				CacheMgt.get().reset(MWarehouse.Table_Name, wh.get_ID());
-			}
-			
-			product = new MProduct(ctx, 0, null);
-			product.setM_Product_Category_ID(DictionaryIDs.M_Product_Category.CHEMICALS.id);
-			product.setName("testSkipProductWithSerial");
-			product.setValue("testSkipProductWithSerial");
-			product.setProductType(MProduct.PRODUCTTYPE_Item);
-			product.setIsStocked(true);
-			product.setIsSold(true);
-			product.setIsPurchased(true);
-			product.setC_UOM_ID(DictionaryIDs.C_UOM.EACH.id);
-			product.setC_TaxCategory_ID(DictionaryIDs.C_TaxCategory.STANDARD.id);
-			product.setM_AttributeSet_ID(DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id);
-			product.saveEx();
-	
-			MPriceListVersion plv = MPriceList.get(DictionaryIDs.M_PriceList.PURCHASE.id).getPriceListVersion(null);
-			MProductPrice pp = new MProductPrice(Env.getCtx(), 0, getTrxName());
-			pp.setM_PriceList_Version_ID(plv.getM_PriceList_Version_ID());
-			pp.setM_Product_ID(product.get_ID());
-			pp.setPriceStd(new BigDecimal("2"));
-			pp.setPriceList(new BigDecimal("2"));
-			pp.saveEx();
-			
-			MAttributeSetInstance asi = new MAttributeSetInstance(Env.getCtx(), 0, getTrxName());
-			asi.setM_AttributeSet_ID(DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id);
-			asi.setSerNo("testSkipProductWithSerial #1");
-			asi.saveEx();
-			
-			createPOAndMRForProduct(product.get_ID(), asi);
-						
-			MStorageOnHand[] onhands = MStorageOnHand.getOfProduct(Env.getCtx(), product.get_ID(), getTrxName());
-			assertEquals(1, onhands.length, "Unexpected number of on hand records");
-			assertEquals(onhands[0].getM_AttributeSetInstance_ID(), asi.get_ID(), "Unexpected M_AttributeSetInstance_ID for on hand record");
-			
-			MAttributeSetInstance asi1 = new MAttributeSetInstance(Env.getCtx(), 0, getTrxName());
-			asi1.setM_AttributeSet_ID(DictionaryIDs.M_AttributeSet.FERTILIZER_LOT.id);
-			asi1.saveEx();
-
-			createPOAndMRForProduct(product.get_ID(), asi1);
-			
-			onhands = MStorageOnHand.getOfProduct(Env.getCtx(), product.get_ID(), getTrxName());
-			assertEquals(2, onhands.length, "Unexpected number of on hand records");
-			Arrays.sort(onhands, new MStorageOnHand(Env.getCtx(), 0, null));
-			assertEquals(onhands[0].getM_AttributeSetInstance_ID(), asi.get_ID(), "Unexpected M_AttributeSetInstance_ID for first on hand record");
-			assertEquals(onhands[1].getM_AttributeSetInstance_ID(), asi1.get_ID(), "Unexpected M_AttributeSetInstance_ID for second on hand record");
-			
-			MInventory inventory = new MInventory(ctx, 0, trxName);
-			inventory.setM_Warehouse_ID(DictionaryIDs.M_Warehouse.HQ.id);
-			inventory.setC_DocType_ID(DictionaryIDs.C_DocType.MATERIAL_PHYSICAL_INVENTORY.id);
-			inventory.saveEx();
-	
-			MInventoryLine iline = new MInventoryLine(inventory,
-					DictionaryIDs.M_Locator.HQ.id, 
-					product.getM_Product_ID(),
-					0, // M_AttributeSetInstance_ID
-					new BigDecimal("2"), // QtyBook
-					new BigDecimal("1"));
-			iline.saveEx();
-	
-			//should success with qty difference being applied to the asi with null serno record
-			ProcessInfo info = MWorkflow.runDocumentActionWorkflow(inventory, DocAction.ACTION_Complete);
-			assertFalse(info.isError(), info.getSummary());
-			inventory.load(getTrxName());
-			assertEquals(DocAction.STATUS_Completed, inventory.getDocStatus());
-			
-			onhands = MStorageOnHand.getOfProduct(Env.getCtx(), product.get_ID(), getTrxName());
-			assertEquals(2, onhands.length, "Unexpected number of on hand records");
-			Arrays.sort(onhands, new MStorageOnHand(Env.getCtx(), 0, null));
-			assertEquals(onhands[0].getM_AttributeSetInstance_ID(), asi.get_ID(), "Unexpected M_AttributeSetInstance_ID for first on hand record");
-			assertEquals(onhands[1].getM_AttributeSetInstance_ID(), asi1.get_ID(), "Unexpected M_AttributeSetInstance_ID for second on hand record");
-			assertEquals(1, onhands[0].getQtyOnHand().intValue());
-			assertEquals(0, onhands[1].getQtyOnHand().intValue());
-		} finally {
-			rollback();
-			set.setIsSerNo(false);
-			set.saveEx();
-			
-			if (product != null) 
-				product.deleteEx(true);
-			
-			if (!disallow) {
-				wh.setIsDisallowNegativeInv(false);
-				wh.saveEx();
-				CacheMgt.get().reset(MWarehouse.Table_Name, wh.get_ID());
-			}
 		}
 	}
 }
