+++ org.idempiere.test/src/org/idempiere/test/model/SalesOrderTest.java
@@ -69,10 +69,12 @@ import org.compiere.wf.MWorkflow;
 import org.idempiere.test.AbstractTestCase;
 import org.idempiere.test.DictionaryIDs;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Isolated;
 
 /**
  * @author hengsin
  */
+@Isolated
 public class SalesOrderTest extends AbstractTestCase {
 
 	public SalesOrderTest() {
